backend/requirements.txt
numpy
flask
pandas
scikit-learn
flask-cors
python-dotenv
azure-cognitiveservices-vision-computervision
pillow
pymongo
certifi
bcrypt
ollama
openai
transformers 
peft 
torch 
safetensors